Updated TenantRoles

+++ src/Services/Api/Controllers/Resources/TenantRolesController.php
@@ -11,6 +11,7 @@ use Bayfront\Bones\Services\Api\Controllers\Abstracts\PrivateApiController;
 use Bayfront\Bones\Services\Api\Controllers\Interfaces\ScopedResourceInterface;
 use Bayfront\Bones\Services\Api\Exceptions\BadRequestException;
 use Bayfront\Bones\Services\Api\Exceptions\ConflictException;
+use Bayfront\Bones\Services\Api\Exceptions\ForbiddenException;
 use Bayfront\Bones\Services\Api\Exceptions\NotFoundException;
 use Bayfront\Bones\Services\Api\Exceptions\UnexpectedApiException;
 use Bayfront\Bones\Services\Api\Models\Resources\TenantRolesModel;
@@ -57,7 +58,14 @@ class TenantRolesController extends PrivateApiController implements ScopedResour
 
         try {
 
-            $id = $this->tenantRolesModel->create($args['tenant_id'], $attrs);
+            if ($this->user->hasAnyPermissions([
+                'global.admin',
+                'tenants.roles.create'
+            ])) {
+                $id = $this->tenantRolesModel->create($args['tenant_id'], $attrs);
+            } else {
+                $id = $this->tenantRolesModel->create($args['tenant_id'], $attrs, $this->filters->doFilter('api.protected.roles', []));
+            }
 
             $created = $this->tenantRolesModel->get($args['tenant_id'], $id);
 
@@ -65,6 +73,8 @@ class TenantRolesController extends PrivateApiController implements ScopedResour
             App::abort(400, $e->getMessage());
         } catch (ConflictException $e) {
             App::abort(409, $e->getMessage());
+        } catch (ForbiddenException $e) {
+            App::abort(403, $e->getMessage());
         } catch (NotFoundException $e) {
             App::abort(404, $e->getMessage());
         }
@@ -185,7 +195,14 @@ class TenantRolesController extends PrivateApiController implements ScopedResour
 
         try {
 
-            $this->tenantRolesModel->update($args['tenant_id'], $args['role_id'], $attrs);
+            if ($this->user->hasAnyPermissions([
+                'global.admin',
+                'tenants.roles.update'
+            ])) {
+                $this->tenantRolesModel->update($args['tenant_id'], $args['role_id'], $attrs);
+            } else {
+                $this->tenantRolesModel->update($args['tenant_id'], $args['role_id'], $attrs, $this->filters->doFilter('api.protected.roles', []));
+            }
 
             $updated = $this->tenantRolesModel->get($args['tenant_id'], $args['role_id']);
 
@@ -193,6 +210,8 @@ class TenantRolesController extends PrivateApiController implements ScopedResour
             App::abort(400, $e->getMessage());
         } catch (ConflictException $e) {
             App::abort(409, $e->getMessage());
+        } catch (ForbiddenException $e) {
+            App::abort(403, $e->getMessage());
         } catch (NotFoundException $e) {
             App::abort(404, $e->getMessage());
         }
@@ -213,6 +232,7 @@ class TenantRolesController extends PrivateApiController implements ScopedResour
      * @throws ContainerNotFoundException
      * @throws HttpException
      * @throws InvalidStatusCodeException
+     * @throws UnexpectedApiException
      */
     public function delete(array $args): void
     {
@@ -225,10 +245,19 @@ class TenantRolesController extends PrivateApiController implements ScopedResour
 
         try {
 
-            $this->tenantRolesModel->delete($args['tenant_id'], $args['role_id']);
+            if ($this->user->hasAnyPermissions([
+                'global.admin',
+                'tenants.roles.delete'
+            ])) {
+                $this->tenantRolesModel->delete($args['tenant_id'], $args['role_id']);
+            } else {
+                $this->tenantRolesModel->delete($args['tenant_id'], $args['role_id'], $this->filters->doFilter('api.protected.roles', []));
+            }
 
             $this->response->setStatusCode(204)->send();
 
+        } catch (ForbiddenException $e) {
+            App::abort(403, $e->getMessage());
         } catch (NotFoundException $e) {
             App::abort(404, $e->getMessage());
         }
